# docs/swagger.json
{
	"components": {
...
		"schemas": {
			"HTTPError": {
				"properties": {
					"code": {
						"type": "number",
						"format": "double"
					},
					"details": {
						"type": "string"
					},
					"additional_info": {}
				},
				"required": [
					"code",
					"details"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Result": {
				"properties": {
					"result": {},
					"error": {
						"$ref": "#/components/schemas/HTTPError"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"errorMsg": {
				"properties": {
					"message": {
						"type": "string"
					},
					"additional_info": {}
				},
				"required": [
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IEvent": {
				"properties": {
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"startedAt": {
						"type": "string"
					},
					"endedAt": {
						"type": "string"
					},
					"communityId": {
						"type": "number",
						"format": "double"
					},
					"isPublic": {
						"type": "boolean"
					},
					"imageId": {
						"type": "number",
						"format": "double"
					},
					"type": {
						"type": "string"
					},
					"site": {
						"type": "string"
					},
					"capacity": {
						"type": "number",
						"format": "double"
					},
					"additionalInfo": {
						"type": "string"
					}
				},
				"required": [
					"title",
					"startedAt",
					"isPublic",
					"type",
					"site",
					"capacity"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ICommunity": {
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"locationId": {
						"type": "number",
						"format": "double"
					},
					"imageId": {
						"type": "number",
						"format": "double"
					},
					"bannerImageId": {
						"type": "number",
						"format": "double"
					},
					"tagIds": {
						"items": {
							"type": "number",
							"format": "double"
						},
						"type": "array"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"accessToken": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "backend",
		"version": "1.0.0",
		"license": {
			"name": "MIT"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/communities/{CID}/events": {
			"post": {
				"operationId": "CreateEvent",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Result"
								},
								"examples": {
									"Example 1": {
										"value": {
											"data": {
												"id": 1
											}
										}
									}
								}
							}
						}
					},
					"403": {
						"description": "Unathorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/errorMsg"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "You are not an admin of this community"
										}
									}
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/errorMsg"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Community not found"
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/errorMsg"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Failed to get community"
										}
									}
								}
							}
						}
					}
				},
				"description": "Create events in a community",
				"tags": [
					"Community"
				],
				"security": [],
				"parameters": [
					{
						"description": "Community ID",
						"in": "path",
						"name": "CID",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IEvent"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetEvents",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Result"
								},
								"examples": {
									"Example 1": {
										"value": {
											"data": {
												"id": 1,
												"name": "Komunitas IT",
												"slug": null,
												"description": "Komunitas IT",
												"createdAt": "2022-04-01T20:07:48.687Z",
												"updatedAt": "2022-04-01T20:07:48.687Z",
												"events": []
											}
										}
									}
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/errorMsg"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Community not found"
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/errorMsg"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Failed to get community"
										}
									}
								}
							}
						}
					}
				},
				"description": "Get events in a community",
				"tags": [
					"Community"
				],
				"security": [],
				"parameters": [
					{
						"description": "Community ID",
						"in": "path",
						"name": "CID",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "true to get upcoming events or false to get past events",
						"in": "query",
						"name": "upcoming",
						"required": true,
						"schema": {
							"type": "boolean"
						}
					}
				]
			}
		},
		"/communities/{CID}": {
			"get": {
				"operationId": "GetCommunityDetail",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Result"
								},
								"examples": {
									"Example 1": {
										"value": {
											"data": {
												"id": 1,
												"name": "Komunitas IT",
												"slug": "komunitas-it",
												"description": "Komunitas hantu bawa balon",
												"locationId": 1,
												"imageId": 1,
												"createdAt": "2022-03-18T14:34:40.159Z",
												"updatedAt": "2022-03-18T14:34:40.159Z"
											}
										}
									}
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/errorMsg"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Community not found"
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/errorMsg"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Could not get community detail"
										}
									}
								}
							}
						}
					}
				},
				"description": "Retrieves a community's detail",
				"tags": [
					"Community"
				],
				"security": [
					{
						"accessToken": []
					}
				],
				"parameters": [
					{
						"description": "The community's ID",
						"in": "path",
						"name": "CID",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Result"
								},
								"examples": {
									"Example 1": {
										"value": {
											"data": {
												"name": "communitree"
											}
										}
									}
								}
							}
						}
					},
					"403": {
						"description": "Forbidden",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/errorMsg"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "You are not admin of this community"
										}
									}
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/errorMsg"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Community not found"
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/errorMsg"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Failed to update community"
										}
									}
								}
							}
						}
					}
				},
				"description": "Update a communitywith the corresponding ID",
				"tags": [
					"Community"
				],
				"security": [
					{
						"accessToken": []
					}
				],
				"parameters": [
					{
						"description": "The community's ID",
						"in": "path",
						"name": "CID",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ICommunity"
							}
						}
					}
				}
			}
		},
		"/communities/{CID}/members": {
			"get": {
				"operationId": "GetMembersId",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Result"
								},
								"examples": {
									"Example 1": {
										"value": {
											"data": [
												{
													"id": 1
												}
											]
										}
									}
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/errorMsg"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Community not found"
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/errorMsg"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Could not get community members"
										}
									}
								}
							}
						}
					}
				},
				"description": "Retrieves a community's members",
				"tags": [
					"Community"
				],
				"security": [
					{
						"accessToken": []
					}
				],
				"parameters": [
					{
						"description": "The community's ID",
						"in": "path",
						"name": "CID",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/communities/{CID}/members/count": {
			"get": {
				"operationId": "GetMembersCount",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Result"
								},
								"examples": {
									"Example 1": {
										"value": {
											"data": 9
										}
									}
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/errorMsg"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Community not found"
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/errorMsg"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Could not get community members count"
										}
									}
								}
							}
						}
					}
				},
				"description": "Retrieves a community's members count",
				"tags": [
					"Community"
				],
				"security": [],
				"parameters": [
					{
						"description": "The community's ID",
						"in": "path",
						"name": "CID",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/communities": {
			"post": {
				"operationId": "Create",
				"responses": {
					"201": {
						"description": "Successfully created community",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Result"
								},
								"examples": {
									"Example 1": {
										"value": {
											"data": "communitree"
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/errorMsg"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Failed to create community"
										}
									}
								}
							}
						}
					}
				},
				"description": "Creates a new community and set the creator as admin",
				"tags": [
					"Community"
				],
				"security": [
					{
						"accessToken": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ICommunity"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetCommunities",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Result"
								},
								"examples": {
									"Example 1": {
										"value": {
											"data": [
												{
													"id": 20,
													"name": "Komunitas IT",
													"slug": null,
													"description": "Komunitas IT",
													"createdAt": "2022-04-01T18:36:39.146Z",
													"updatedAt": "2022-04-01T18:36:39.146Z",
													"hasTags": [],
													"locatedOn": null,
													"hasImage": null
												},
												{
													"id": 21,
													"name": "Komunitas Otomotif",
													"slug": null,
													"description": "Komunitas ",
													"createdAt": "2022-04-01T18:36:39.146Z",
													"updatedAt": "2022-04-01T18:36:39.146Z",
													"hasTags": [],
													"locatedOn": {
														"city": "Jakarta",
														"stateProvince": "DKI Jakarta",
														"country": "Indonesia",
														"postalCode": "40136"
													},
													"hasImage": {
														"url": "./foto.png"
													}
												}
											]
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/errorMsg"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Failed to retrieve communities"
										}
									}
								}
							}
						}
					}
				},
				"description": "Get all communities",
				"tags": [
					"Community"
				],
				"security": [],
				"parameters": []
			}
		},
		"/communities/{CID}/members/status": {
			"get": {
				"operationId": "GetMembershipStatus",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Result"
								},
								"examples": {
									"Example 1": {
										"value": {
											"data": {
												"status": "member"
											}
										}
									}
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/errorMsg"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "User not authorized"
										}
									}
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/errorMsg"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Community not found"
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/errorMsg"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Failed to get community"
										}
									}
								}
							}
						}
					}
				},
				"description": "Get user's membership status in a community",
				"tags": [
					"Community"
				],
				"security": [
					{
						"accessToken": []
					}
				],
				"parameters": [
					{
						"description": "id of the community",
						"in": "path",
						"name": "CID",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/api/v1"
		}
	]
}